+++ packages/e2e-tests/plugins/interactive-blocks/router-navigate/render.php
@@ -12,7 +12,7 @@ wp_enqueue_script_module( 'router-navigate-view' );
 if ( $attributes['disableNavigation'] ) {
 	wp_interactivity_config(
 		'core/router',
-		array( 'clientNavigation' => false )
+		array( 'clientNavigationDisabled' => true )
 	);
 }
 ?>