+++ packages/interactivity-router/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### New Features
 
--   Add the `clientNavigation` option to the `core/router` config. ([58749](https://github.com/WordPress/gutenberg/pull/58749))
+-   Add the `clientNavigationDisabled` option to the `core/router` config. ([58749](https://github.com/WordPress/gutenberg/pull/58749))
 
 ## 1.0.0 (2024-01-24)